refactor: move stuff to an `exchange` module
