+++ src/exchange/data-access/exchange-dao.js
@@ -1,4 +1,4 @@
-import data from '../../data.json'
+import data from '../../../data.json'
 
 export const getAllExchanges = async () => data.exchanges
